feat(no-products-found): extracted ProductCard component
