--- packages/falcon-ecommerce-uikit/src/ProductsList/ProductCardLayout.tsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import { themed } from '@deity/falcon-ui';
-
-export const ProductCardLayout = themed({
-  tag: Link,
-  defaultTheme: {
-    card: {
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'center',
-      color: 'secondaryText',
-
-      css: {
-        height: '100%',
-        textDecoration: 'none',
-        overflow: 'hidden',
-        cursor: 'pointer'
-      }
-    }
-  }
-});
