+++ packages/falcon-ecommerce-uikit/src/ProductsList/index.tsx
@@ -1,3 +1,3 @@
 export * from './ProductsList';
 export * from './ProductsListQuery';
-export * from './ProductCardLayout';
+export * from './ProductCard';
